Remove memoization of events in _IonNature. (#97)
Remove memoization of eents in _IonNature.

This memoization was causing #95 because the IonEvent.field_name field
was being used when writing out struct fields instead of the key of the
field within IonPyDict.  This meant that field with a value that
originated from another IonPyDict to retain field name from the original
IonPyDict.