Remove Python 3.6 from Builds, Bump Minor Version
This change move the oldest supported version of Python to 3.7.
Python 3.6 is now a year past EOL and is no longer supported on the
ubuntu images used for github builds.

This also bumps the *minor* version number. While there are no current
changes that _shouldn't_ work on 3.6, future changes may not.