Define IonPyBool as Distinct from IonPyInt
Previously Ion Bools were modeled as Ion Ints of either 0 or 1 with
the IonType set to BOOL. This worked correctly so long as folks were
looking at the IonType, or using simpleion.dump() which did so.

It was confusing however. The representation of an Ion Bool was as a
0 or 1, and the type() of instances could be IonPyInt. Code which used
the Ion class type for branching could also be incorrect.

Consider:
```
>>> b = simpleion.loads('true')
>>> b
1
>>> type(b)
<class 'amazon.ion.simple_types.IonPyInt'>
>>> b.ion_type
<IonType.BOOL: 1>
>>> simpleion.dumps(b, binary=False)
'$ion_1_0 true'
```

This change adds a specialization of _IonNature to provide more
intuitive behavior.

Now:
```
>>> b = simpleion.loads('true')
>>> b
<amazon.ion.simple_types.IonPyBool object at 0x107408130>
>>> type(b)
<class 'amazon.ion.simple_types.IonPyBool'>
>>> b.ion_type
<IonType.BOOL: 1>
>>> simpleion.dumps(b, binary=False)
'$ion_1_0 true'
```

While the __repr__ has changed to provide less apparent information it
is less confusing. It is also worth noting that other values (such as
null) show similarly and annotations don't show up, so there should be
no expectation that __repr__ will show semantically meaningful Ion data.