Replace Usages of record with NamedTuple (#262)
* Replace Usages of record with NamedTuple

This change replaces all of the usages of the record super-type
constructor with the built-in NamedTuple. This makes the code easier
to read (no custom special type mojo) and enables IDEs and other tooling
to provide type hints when constructing and field name completion when
using instances.

The types which extend NamedTuple are immutable as they were before.
I'm sure there was a reason record was created initially, but I don't
see a good reason to keep it now, particularly given the benefits of
just using NamedTuple.

I kept the record constructor and metaclass to avoid breaking a
consumer who is using them. We can remove them when we release an MV.

* Correct Type Hint for Transition event