Warn on ImportError when importing ionc
Before this change ImportErrors other than ModuleNotFoundErrors would
propagate. Unless a user was catching the importing of the simpleion
module (and why would they?) this would result in their process
crashing. Users may have encountered this in platforms that disallow
c extensions, such as spark, or if they had an ionc dynamic library
that didn't match the architecture of the python runtime.

Now it will warn but import will proceed with pure python mode.