Add repo to PR head checkout
In order to support 3rd party repos it looks like we can either leave both the
revision, and the repo off of the config, or specify both. By not specifying
the repo, but specifying the ref/revision, the checkout action will try to pull
the revision from the target repo.