Initial CMakeLists.txt Files
This commit adds CMakeLists.txt files for the overall ion-python
project and the python c-extension.

It allows one to build the c-extension as a  project (assuming
submodules are init'ed and updated) using cmake. Which is what CLion
expects. What is built is not currently loadable by a Python Interpreter.

On it's own this isn't really impactful. But it could be used as part
of a setup for a "Python Interpreter Free" test/debug/profiling setup.

Or... there may be merit in using something like py-build-cmake to
make the project PEP 517 compliant with a pyproject.toml file. I
haven't looked to hard at that, but I think there is an opportunity
there.

I had to add prefixes to the imports in the _ioncmodule.h file to make
it work. That may or may not be the right answer.