Avoid strcpy of Field Names in C-Ext Read
This change removes a strcpy to a temporary stack allocated array for
the field name of a Struct field. Instead it just allocates the
PyUnicode object for the field name, which does a copy anyway. The
copy is important because the ION_STRING is just a pointer to the
buffer which will change as we descend into parsing the Struct.

The performance impact of this change is measurable for some datasets
but not terribly significant. It is nonetheless an apparent simplifier
and therefore worth doing regardless.