Simplify ION_DECIMAL to Python Decimal Conversion
This change simplifies the conversion of an ION_DECIMAL to a Python
Decimal when reading Decimals. Instead of converting the ION_DECIMAL
to a tuple of (sign, (digit, ...), exponent) it just gets the string
representation and uses the corresponding Decimal constructor.

This turns out to be about 20% faster for marshalling "bare value"
decimals of reasonable size (think amounts of money) and uses less
memory.

I also removed an unnecessary extra byte padding for int conversion,
removed a superfluous error check and clarified with comments.