Reorder simpleion and Refactor Pydoc
This change puts the user API methods at the top of the file
and moves the pydoc to the actual API methods. I followed the std
json module in the choice to put the main doc strings on dump and
load and have dumps and loads refer to them.

It also factors out the tables that were in loads and dumps into
a single table. I found the previous factoring very difficult to grok
and I believe this is more readable.

I also removed all of the unused kwarg parameters and their pydoc
strings. It added a lot of cognitive load for no value.

Finally I made some minor cleanings to Arg descs and corrected the
default text buffer value for loading with the C-Extension.