Try to read and convert Ion INT as int 64
This change makes it so that we try to read Ion INTs as int64s and
convert them to Python with PyLong_FromLongLong. If the int is too
large then we fallback to converting via base 10 string repr.

I don't like adding complexity but this is significantly faster for
datasets with small ints. For example, the "service_log_legacy"
benchmark is about 7% faster with this change.