Implement Symbol as Text in C-extension
By default the IonPy value for a Symbol is IonPySymbol and the bare
value is a SymbolToken. With the SYMBOL_AS_TEXT flag set, the IonPy
value for a Symbol is IonPyText, and the bare value is str.

This makes Symbol handling simpler if one only cares about the text
value. It also improves load performance for symbols significantly.

Symbols with undefined text cannot be emitted with this flag set and
will raise exceptions.