Build Std Dict without Lists (#331)
* Build Std Dict without Lists

The previous STRUCT_AS_STD_DICT commit still builds a "hashmap of
lists" structure for "std" dicts, which was not intended. It was a
failure to get my WIP/POC code into that PR, which this change remedies.

I added a new test to actually verify the behavior. I kept but did
not add to the existing test_value_model_flags because that has a role
in covering the breadth of the IonPyValueFlags and I didn't want to
overcomplicate that for this Struct specific behavior.

* Update amazon/ion/ioncmodule.c

Co-authored-by: Tyler Gregg <tgregg@users.noreply.github.com>

---------

Co-authored-by: Tyler Gregg <tgregg@users.noreply.github.com>