Fix Bug with Ion Text<>Binary Conversion
This change fixes a bug in the Ion Text<>Binary conversion where
an Ion stream of multiple top-level values would be transcoded as
an Ion list.

In the process of fixing it I cleaned up all of the Format comparisons
to be identity of enum constant instead of string comparison of value.
This is a lot more succinct and avoid helps to avoid typo errors.