Use PyDateTime C API to Write Timestamps
This change uses the PyDateTime C API to access the Timestamp
attributes from the datetime super class. It improves Timestamp
serialization by about 10% for both Text and Binary. I expect the
improvement for basic datetimes to be higher but I haven't measured
it.

It should be compatible with future changes to the Timestamp types
so long as they still inherit from DateTime, which I think is a safe
assumption.