Use PyDateTime C API to Write Timestamps (#336)
* Use PyDateTime C API to Write Timestamps

This change uses the PyDateTime C API to access the Timestamp
attributes from the datetime super class. It improves Timestamp
serialization by about 10% for both Text and Binary. I expect the
improvement for basic datetimes to be higher but I haven't measured
it.

It should be compatible with future changes to the Timestamp types
so long as they still inherit from DateTime, which I think is a safe
assumption.

* Remove pip from requirements.txt

The GH windows build is now failing because we are trying to load
a different version of pip than is _the one actually running_.

I do not know why pip was added explicitly here but it is sort of
nonsensical: how can you install the requirements with pip if you
don't already have... pip?