Refactor Timestamp construction
This change pulls the heavy validation logic out of the __new__ and
into the __init__. This preserves the contract for the constructor
while providing a "fast" object construction for those who call
__new__ directly.

In the next commit, I change the ionc extension module to call it.

I also added the fractional_seconds to __repr__ because it was
difficult to debug without that.