Refactor simpleion pure-python load methods
This change removes some duplication in the simpleion implementation.

Before this change the _load and _load_iteratively methods were near
duplicates and used the python call stack for maintaining the Ion
container stack.

After this change _load is removed in favor of _load_iteratively which
uses a deque to maintain the container stack.

It also factors out the handling of the single_value and parse_eagerly
flags which was also nearly duplicated between the ionc and
pure-python usages.