Working Parsing of Bools, Null (untyped), and Whitespace
Some fairly savage hacking at both the SliceableBuffer and protons
abstractions. It works and I _think_ it's the right contracts for the
components but not totally sure.

I think at a minimum I need to fix the EOF marking so that it is
clearly correct for all cases. That means an actual flag. But I don't
want to have the parsers or the buffer have to track that, it should
just be in the parser. I need to track depth as context anyway, so
we can have the parse methods in the reader track that as well.