Changes for Text Parser Refactor
I'm putting this on the shelf for a bit.

Basically: I did some stuff to move this along, I ran it against my
own "hkc" dataset and the results were not great.

Specifically: the existing pure-python text parser does about .3
ops/second whereas this code does .4. And it's not complete.

Struct parsing is broken becuase as of now each of the container
parsers is stateless, but the parsers themselves are not re-entrant.
So you effectively can't have containers inside of containers.

It took me about 3 - 4 heads down days to get here and I suspect it
would take about as much to productionize it.

Given what I saw from messing around, simply changing out the main
value alt for a table (or reading only once and passing the data?)
would likely have a significant impact on performance.

Changes:
* Added up to day precision Timestamp parsing
* Broken Semantically, but "working" Decimal Parsing
* Untested formally but attempted "table" proton
* Hacked up ability to run Refactor against tests