feat: Add `trailing_commas` to `dump`/`dumps`
Add an optional boolean argument `trailing_commas` to the `dump`
and `dumps` methods in `simpleion`. Default: False.

When `trailing_commas=True`, and when `indent` is not `None`, this
causes `dump[s]` to include a comma on the last item in a container.

**Note:** Since the C module already **does not support pretty
printing**, this flag does nothing when using the C module.